feat: Subscribe topics when user enable push notification
