+++ lib/l10n/localization.dart
@@ -3,7 +3,9 @@ import 'package:intl/intl.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'localization.g.dart';
+
 part 'localization_en.dart';
+
 part 'localization_ja.dart';
 
 /// A provider that returns the current [Localization].
@@ -115,6 +117,8 @@ abstract base class Localization {
   String get settingsPushNotificationPermanentlyDenied;
 
   String get settingsPushNotificationMessageAuthorized;
+
+  String get settingsPushNotificationMessageAlreadyAuthorized;
 
   String get settingsPushNotificationMessageDenied;
