+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:conference_2023/app.dart';
-import 'package:conference_2023/model/messaging.dart';
 import 'package:conference_2023/model/shared_preferences.dart';
 import 'package:conference_2023/util/font_license.dart';
 import 'package:conference_2023/util/provider_state_logger.dart';
@@ -32,9 +31,6 @@ void main() async {
     FirebaseCrashlytics.instance.recordError(error, stack);
     return true;
   };
-
-  /// initialize firebase apps
-  await initFirebaseMessaging();
 
   fontLicenses();
